requirements.txt
fastapi
pandas
pydantic
uvicorn
python-multipart
python-dotenv
openai-agents
langchain-openai
langchain-chroma
langchain-community
langchain-core
ipykernel
langchain-unstructured
unstructured-client
unstructured[pdf]
python-magic
chromadb
langchain-experimental
pypdf
streamlit-mic-recorder
streamlit
streamlit-chat